Second pass:trying to bring front page in shape
Adding all the content into the jumbotrons was imposible.
This is a humble attempt (i should not do this) to edit the larger text into smaller chunks.
Includes also new media breakpoints for the jumbotrons and credits for the main image.
A lot missing still. We need to move this out of the static template. This is just for our internal use and makes little sense in custom deployments.